+++ HR.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HR.LeaveManagement.Application.Persistense.Contracts;
 using MediatR;
@@ -18,6 +19,10 @@ namespace HR.LeaveManagement.Application.Features.LeaveRequests.Handlers.Command
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await leaveRequestRepository.Get(request.Id);
+            if (leaveRequest is null)
+            {
+                throw new NotFoundException(nameof(leaveRequest), request.Id);
+            }
             await leaveRequestRepository.Delete(leaveRequest);
             return Unit.Value;
         }
